requirements.txt
fastapi[standard]
sqlalchemy
aiosqlite
PyJWT
